requirements.txt
apache-airflow-providers-google>=10.0.0
google-cloud-storage>=2.13.0
google-cloud-bigquery>=3.17.2
pandas>=2.2.0
pyarrow>=15.0.0
great-expectations>=0.18.21
dbt-bigquery>=1.7.4
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dateutil==2.8.2
pytz==2024.1
